Moved predef.hpp to detail
