+++ include/boost/simd/arch/arm/neon/spec.hpp
@@ -14,7 +14,7 @@
 #ifndef BOOST_SIMD_ARCH_ARM_NEON_SPEC_HPP_INCLUDED
 #define BOOST_SIMD_ARCH_ARM_NEON_SPEC_HPP_INCLUDED
 
-#include <boost/simd/sdk/predef.hpp>
+#include <boost/simd/detail/predef.hpp>
 
 #if BOOST_HW_SIMD_ARM == BOOST_HW_SIMD_ARM_NEON_VERSION
